requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
